Propagate write errors
