+++ api/src/write.rs
@@ -1,4 +1,19 @@
 use core::ffi::c_void;
+use shopify_function_wasm_api_core::write::WriteResult;
+
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    #[error("I/O error")]
+    IoError,
+}
+
+fn map_result(result: u32) -> Result<(), Error> {
+    match WriteResult::from_repr(result) {
+        Some(WriteResult::Ok) => Ok(()),
+        Some(WriteResult::IoError) => Err(Error::IoError),
+        None => panic!("Unknown write result: {}", result),
+    }
+}
 
 pub struct ValueSerializer(*mut c_void);
 
@@ -7,12 +22,12 @@ impl ValueSerializer {
         Self(unsafe { crate::shopify_function_output_new() as *mut _ })
     }
 
-    pub fn write_bool(&mut self, value: bool) {
-        unsafe { crate::shopify_function_output_new_bool(self.0 as _, value as u32) };
+    pub fn write_bool(&mut self, value: bool) -> Result<(), Error> {
+        map_result(unsafe { crate::shopify_function_output_new_bool(self.0 as _, value as u32) })
     }
 
-    pub fn finalize(&mut self) {
-        unsafe { crate::shopify_function_output_finalize(self.0 as _) };
+    pub fn finalize(&mut self) -> Result<(), Error> {
+        map_result(unsafe { crate::shopify_function_output_finalize(self.0 as _) })
     }
 }
